karel

+++ BE/Configuration.cs
@@ -8,8 +8,8 @@ namespace BE
 {
     public static class Configuration
     {
-        public static int NumStaticGuestRequest = 10000000;
-        public static int NumStaticHostingUnit = 10000000;
+        public static long NumStaticGuestRequest = 10000000;
+        public static long NumStaticHostingUnit = 10000000;
         public static int NumStaticOrder = 1;
         public static int Commission = 10;
         public static string WEB_LICENSE_MAIL;
