+++ BL/IBL.cs
@@ -13,7 +13,7 @@ namespace BL
         #region GuestRequestFonction
         void AddGuestRequest(GuestRequest gs);
         void UpdateGuestRequest(GuestRequest gs);
-        GuestRequest GetGuestRequest(int key);
+        GuestRequest GetGuestRequest(long key);
         #endregion
 
         #region HostFonction
@@ -25,9 +25,9 @@ namespace BL
 
         #region HostingUnitFonction
         void AddHostingUnit(HostingUnit hu);
-        void EraseHostingUnit(int Key);
+        void EraseHostingUnit(long Key);
         void UpdateHostingUnit(HostingUnit hu);
-        HostingUnit GetHostingUnit(int key);
+        HostingUnit GetHostingUnit(long key);
         #endregion
 
         #region OrderFonction
